+++ lib/segment/benches/sparse_index_search.rs
@@ -9,7 +9,10 @@ use indicatif::{ProgressBar, ProgressDrawTarget, ProgressStyle};
 use itertools::Itertools as _;
 use rand::SeedableRng;
 use rand::rngs::StdRng;
-use segment::fixtures::sparse_fixtures::fixture_sparse_index_from_iter;
+use segment::fixtures::sparse_fixtures::{
+    fixture_sparse_index_from_iter, mmap_from_ram, mmap_open, open_sparse_index, ram_from_ram,
+    ram_open,
+};
 use segment::index::sparse_index::sparse_index_config::{SparseIndexConfig, SparseIndexType};
 use segment::index::sparse_index::sparse_vector_index::{
     SparseVectorIndex, SparseVectorIndexOpenArgs,
@@ -79,6 +82,8 @@ fn sparse_vector_index_search_benchmark_impl(
         progress("Indexing (1/2)", vectors_len).wrap_iter(vectors),
         FULL_SCAN_THRESHOLD,
         SparseIndexType::MutableRam,
+        ram_open,
+        ram_from_ram,
     )
     .unwrap();
 
@@ -106,15 +111,19 @@ fn sparse_vector_index_search_benchmark_impl(
         SparseIndexConfig::new(Some(FULL_SCAN_THRESHOLD), SparseIndexType::Mmap, None);
     let pb = progress("Indexing (2/2)", vectors_len);
     let sparse_vector_index_mmap: SparseVectorIndex<InvertedIndexCompressedMmap<f32, MmapFile>> =
-        SparseVectorIndex::open(SparseVectorIndexOpenArgs {
-            config: sparse_index_config,
-            id_tracker: sparse_vector_index.id_tracker().clone(),
-            vector_storage: sparse_vector_index.vector_storage().clone(),
-            payload_index: sparse_vector_index.payload_index().clone(),
-            path: mmap_index_dir.path(),
-            stopped: &stopped,
-            tick_progress: || pb.inc(1),
-        })
+        open_sparse_index(
+            SparseVectorIndexOpenArgs {
+                config: sparse_index_config,
+                id_tracker: sparse_vector_index.id_tracker().clone(),
+                vector_storage: sparse_vector_index.vector_storage().clone(),
+                payload_index: sparse_vector_index.payload_index().clone(),
+                path: mmap_index_dir.path(),
+                stopped: &stopped,
+                tick_progress: || pb.inc(1),
+            },
+            mmap_open::<f32>,
+            mmap_from_ram::<f32>,
+        )
         .unwrap();
     pb.finish_and_clear();
     assert_eq!(sparse_vector_index_mmap.indexed_vector_count(), vectors_len);
