+++ lib/segment/src/fixtures/sparse_fixtures.rs
@@ -1,3 +1,4 @@
+use std::borrow::Cow;
 use std::fmt::Debug;
 use std::path::Path;
 use std::sync::Arc;
@@ -6,29 +7,128 @@ use std::sync::atomic::AtomicBool;
 use atomic_refcell::AtomicRefCell;
 use common::counter::hardware_counter::HardwareCounterCell;
 use common::types::PointOffsetType;
+use common::universal_io::{MmapFile, MmapFs};
 use rand::Rng;
 use sparse::common::sparse_vector::SparseVector;
 use sparse::common::sparse_vector_fixture::random_sparse_vector;
+use sparse::common::types::Weight;
 use sparse::index::inverted_index::InvertedIndex;
+use sparse::index::inverted_index::inverted_index_compressed_immutable_ram::InvertedIndexCompressedImmutableRam;
+use sparse::index::inverted_index::inverted_index_compressed_mmap::InvertedIndexCompressedMmap;
+use sparse::index::inverted_index::inverted_index_ram::InvertedIndexRam;
 
 use crate::common::operation_error::OperationResult;
 use crate::fixtures::payload_context_fixture::create_id_tracker_fixture;
 use crate::index::VectorIndexRead;
 use crate::index::sparse_index::sparse_index_config::{SparseIndexConfig, SparseIndexType};
 use crate::index::sparse_index::sparse_vector_index::{
-    SparseVectorIndex, SparseVectorIndexOpenArgs,
+    SparseOpenPlan, SparseVectorIndex, SparseVectorIndexOpenArgs,
 };
 use crate::index::struct_payload_index::StructPayloadIndex;
 use crate::payload_storage::in_memory_payload_storage::InMemoryPayloadStorage;
 use crate::vector_storage::sparse::mmap_sparse_vector_storage::MmapSparseVectorStorage;
 use crate::vector_storage::{VectorStorage, VectorStorageEnum, VectorStorageRead};
 
-/// Prepares a sparse vector index with a given iterator of sparse vectors
+// Per-type constructors over the local mmap filesystem, passed to the fixtures
+// and `SparseVectorIndex::open` in place of trait-level construction.
+pub fn ram_open(path: &Path) -> common::universal_io::Result<InvertedIndexRam> {
+    InvertedIndexRam::open(&MmapFs, path)
+}
+pub fn ram_from_ram(
+    ram: Cow<InvertedIndexRam>,
+    path: &Path,
+) -> common::universal_io::Result<InvertedIndexRam> {
+    InvertedIndexRam::from_ram_index(&MmapFs, ram, path)
+}
+pub fn iram_open<W: Weight + 'static>(
+    path: &Path,
+) -> common::universal_io::Result<InvertedIndexCompressedImmutableRam<W>> {
+    InvertedIndexCompressedImmutableRam::open(&MmapFs, path)
+}
+pub fn iram_from_ram<W: Weight + 'static>(
+    ram: Cow<InvertedIndexRam>,
+    path: &Path,
+) -> common::universal_io::Result<InvertedIndexCompressedImmutableRam<W>> {
+    InvertedIndexCompressedImmutableRam::from_ram_index(&MmapFs, ram, path)
+}
+pub fn mmap_open<W: Weight + 'static>(
+    path: &Path,
+) -> common::universal_io::Result<InvertedIndexCompressedMmap<W, MmapFile>> {
+    InvertedIndexCompressedMmap::open(&MmapFs, path)
+}
+pub fn mmap_from_ram<W: Weight + 'static>(
+    ram: Cow<InvertedIndexRam>,
+    path: &Path,
+) -> common::universal_io::Result<InvertedIndexCompressedMmap<W, MmapFile>> {
+    InvertedIndexCompressedMmap::from_ram_index(&MmapFs, ram, path)
+}
+
+/// Test/bench convenience that runs the load-or-build orchestration and
+/// assembles a [`SparseVectorIndex`], given the concrete inverted-index
+/// constructors. Production code (`create_sparse_vector_index`) inlines the same
+/// `plan` → construct → `finish` steps per type without these callbacks.
+pub fn open_sparse_index<I: InvertedIndex>(
+    args: SparseVectorIndexOpenArgs<impl FnMut()>,
+    open_index: impl FnOnce(&Path) -> common::universal_io::Result<I>,
+    from_ram_index: impl FnOnce(Cow<InvertedIndexRam>, &Path) -> common::universal_io::Result<I>,
+) -> OperationResult<SparseVectorIndex<I>> {
+    let SparseVectorIndexOpenArgs {
+        config,
+        id_tracker,
+        vector_storage,
+        payload_index,
+        path,
+        stopped,
+        mut tick_progress,
+    } = args;
+    let plan = SparseVectorIndex::<I>::plan(
+        config,
+        &id_tracker,
+        &vector_storage,
+        path,
+        stopped,
+        &mut tick_progress,
+    )?;
+    let (inverted_index, config, indices_tracker, persist) = match plan {
+        SparseOpenPlan::Load {
+            config,
+            indices_tracker,
+        } => (open_index(path)?, config, indices_tracker, false),
+        SparseOpenPlan::Build {
+            config,
+            ram_index,
+            indices_tracker,
+            persist,
+        } => (
+            from_ram_index(Cow::Owned(ram_index), path)?,
+            config,
+            indices_tracker,
+            persist,
+        ),
+    };
+    SparseVectorIndex::finish(
+        config,
+        id_tracker,
+        vector_storage,
+        payload_index,
+        path,
+        inverted_index,
+        indices_tracker,
+        persist,
+    )
+}
+
+/// Prepares a sparse vector index with a given iterator of sparse vectors.
+///
+/// `open_index` / `from_ram_index` build the concrete inverted index (the caller
+/// knows the type and filesystem); see `SparseVectorIndex::open`.
 pub fn fixture_sparse_index_from_iter<I: InvertedIndex>(
     data_dir: &Path,
     vectors: impl ExactSizeIterator<Item = SparseVector>,
     full_scan_threshold: usize,
     index_type: SparseIndexType,
+    open_index: impl FnOnce(&Path) -> common::universal_io::Result<I>,
+    from_ram_index: impl FnOnce(Cow<InvertedIndexRam>, &Path) -> common::universal_io::Result<I>,
 ) -> OperationResult<SparseVectorIndex<I>> {
     let stopped = AtomicBool::new(false);
 
@@ -74,8 +174,8 @@ pub fn fixture_sparse_index_from_iter<I: InvertedIndex>(
     );
 
     let sparse_index_config = SparseIndexConfig::new(Some(full_scan_threshold), index_type, None);
-    let sparse_vector_index: SparseVectorIndex<I> =
-        SparseVectorIndex::open(SparseVectorIndexOpenArgs {
+    let sparse_vector_index: SparseVectorIndex<I> = open_sparse_index(
+        SparseVectorIndexOpenArgs {
             config: sparse_index_config,
             id_tracker,
             vector_storage: vector_storage.clone(),
@@ -83,7 +183,10 @@ pub fn fixture_sparse_index_from_iter<I: InvertedIndex>(
             path: index_dir,
             stopped: &stopped,
             tick_progress: || (),
-        })?;
+        },
+        open_index,
+        from_ram_index,
+    )?;
 
     assert_eq!(
         sparse_vector_index.indexed_vector_count(),
@@ -100,27 +203,16 @@ pub fn fixture_sparse_index<I: InvertedIndex + Debug, R: Rng + ?Sized>(
     max_dim: usize,
     full_scan_threshold: usize,
     data_dir: &Path,
+    open_index: impl FnOnce(&Path) -> common::universal_io::Result<I>,
+    from_ram_index: impl FnOnce(Cow<InvertedIndexRam>, &Path) -> common::universal_io::Result<I>,
 ) -> SparseVectorIndex<I> {
     fixture_sparse_index_from_iter(
         data_dir,
         (0..num_vectors).map(|_| random_sparse_vector(rnd, max_dim)),
         full_scan_threshold,
         SparseIndexType::ImmutableRam,
+        open_index,
+        from_ram_index,
     )
     .unwrap()
-}
-
-#[macro_export]
-macro_rules! fixture_for_all_indices {
-    ($test:ident::<_>($($args:tt)*)) => {
-        eprintln!("InvertedIndexCompressedImmutableRam<f32>");
-        $test::<
-            ::sparse::index::inverted_index::inverted_index_compressed_immutable_ram::InvertedIndexCompressedImmutableRam<f32>
-        >($($args)*);
-
-        eprintln!("InvertedIndexCompressedMmap<f32, MmapFile>");
-        $test::<
-            ::sparse::index::inverted_index::inverted_index_compressed_mmap::InvertedIndexCompressedMmap<f32, MmapFile>
-        >($($args)*);
-    };
 }
