+++ lib/segment/src/index/sparse_index/sparse_vector_index.rs
@@ -1,4 +1,3 @@
-use std::borrow::Cow;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::sync::atomic::AtomicBool;
@@ -13,11 +12,12 @@ use fs_err as fs;
 use sparse::SearchScratchPool;
 use sparse::common::sparse_vector::SparseVector;
 use sparse::index::inverted_index::InvertedIndex;
+use sparse::index::inverted_index::inverted_index_ram::InvertedIndexRam;
 use sparse::index::inverted_index::inverted_index_ram_builder::InvertedIndexBuilder;
 
 use self::read_view::{SparseVectorIndexReadView, SparseVectorIndexReadViewEnum};
 use super::indices_tracker::IndicesTracker;
-use crate::common::operation_error::{OperationError, OperationResult, check_process_stopped};
+use crate::common::operation_error::{OperationResult, check_process_stopped};
 use crate::id_tracker::{IdTrackerEnum, IdTrackerRead};
 use crate::index::sparse_index::sparse_index_config::SparseIndexConfig;
 use crate::index::sparse_index::sparse_search_telemetry::SparseSearchesTelemetry;
@@ -76,104 +76,131 @@ pub struct SparseVectorIndexOpenArgs<'a, F: FnMut()> {
     pub tick_progress: F,
 }
 
+/// Outcome of [`SparseVectorIndex::plan`]: whether the persisted index can be
+/// loaded as-is, or must be (re)built from a freshly assembled RAM index.
+///
+/// The RAM index is type-agnostic, so the caller performs the type-specific
+/// construction (`open` / `from_ram_index`) and then [`SparseVectorIndex::finish`].
+pub enum SparseOpenPlan {
+    /// The persisted index is current; load the inverted index from the path.
+    Load {
+        config: SparseIndexConfig,
+        indices_tracker: IndicesTracker,
+    },
+    /// (Re)build: convert `ram_index` into the inverted index. Persist the
+    /// result when `persist` is set (mutable RAM indexes are not persisted).
+    Build {
+        config: SparseIndexConfig,
+        ram_index: InvertedIndexRam,
+        indices_tracker: IndicesTracker,
+        persist: bool,
+    },
+}
+
 impl<TInvertedIndex: InvertedIndex> SparseVectorIndex<TInvertedIndex> {
-    /// Open a sparse vector index at a given path
-    pub fn open<F: FnMut()>(args: SparseVectorIndexOpenArgs<F>) -> OperationResult<Self> {
-        let SparseVectorIndexOpenArgs {
-            config,
-            id_tracker,
-            vector_storage,
-            payload_index,
-            path,
-            stopped,
-            tick_progress,
-        } = args;
-
-        let config_path = SparseIndexConfig::get_config_path(path);
-
-        let (config, inverted_index, indices_tracker) = if !config.index_type.is_persisted() {
-            // RAM mutable case - build inverted index from scratch and use provided config
+    /// Decide whether the on-disk index can be loaded or must be (re)built.
+    ///
+    /// A rebuild assembles the RAM index here (it is type-agnostic); the caller
+    /// then constructs the concrete inverted index from the returned plan and
+    /// hands everything to [`Self::finish`]. This keeps `SparseVectorIndex` free
+    /// of any inverted-index construction callbacks.
+    ///
+    /// A persisted index is reloaded only when its version file (written last as
+    /// a "build completed" marker) matches the current version; otherwise it is
+    /// rebuilt.
+    pub fn plan(
+        config: SparseIndexConfig,
+        id_tracker: &AtomicRefCell<IdTrackerEnum>,
+        vector_storage: &AtomicRefCell<VectorStorageEnum>,
+        path: &Path,
+        stopped: &AtomicBool,
+        tick_progress: impl FnMut(),
+    ) -> OperationResult<SparseOpenPlan> {
+        if !config.index_type.is_persisted() {
+            // RAM mutable case - build from scratch, keep the provided config, do not persist.
             fs::create_dir_all(path)?;
-            let (inverted_index, indices_tracker) = Self::build_inverted_index(
-                &id_tracker,
-                &vector_storage,
-                path,
-                stopped,
-                tick_progress,
-            )?;
-            (config, inverted_index, indices_tracker)
-        } else {
-            Self::try_load(path).or_else(|e| {
-                if fs::exists(path).unwrap_or(true) {
-                    log::warn!("Failed to load {path:?}, rebuilding: {e}");
+            let (ram_index, indices_tracker) =
+                Self::build_ram_index(id_tracker, vector_storage, stopped, tick_progress)?;
+            return Ok(SparseOpenPlan::Build {
+                config,
+                ram_index,
+                indices_tracker,
+                persist: false,
+            });
+        }
 
-                    // Drop index completely.
-                    fs::remove_dir_all(path)?;
-                }
+        let stored_version = TInvertedIndex::Version::load_universal(&MmapFs, path)?;
+        if stored_version == Some(TInvertedIndex::Version::current()) {
+            let config = SparseIndexConfig::load(&SparseIndexConfig::get_config_path(path))?;
+            let indices_tracker = IndicesTracker::open(path)?;
+            return Ok(SparseOpenPlan::Load {
+                config,
+                indices_tracker,
+            });
+        }
 
-                fs::create_dir_all(path)?;
+        if fs::exists(path).unwrap_or(true) {
+            log::warn!(
+                "Sparse index at {path:?} is missing or outdated (found {stored_version:?}, expected {}), rebuilding",
+                TInvertedIndex::Version::current(),
+            );
+            // Drop index completely.
+            fs::remove_dir_all(path)?;
+        }
+        fs::create_dir_all(path)?;
+        let (ram_index, indices_tracker) =
+            Self::build_ram_index(id_tracker, vector_storage, stopped, tick_progress)?;
+        Ok(SparseOpenPlan::Build {
+            config,
+            ram_index,
+            indices_tracker,
+            persist: true,
+        })
+    }
 
-                let (inverted_index, indices_tracker) = Self::build_inverted_index(
-                    &id_tracker,
-                    &vector_storage,
-                    path,
-                    stopped,
-                    tick_progress,
-                )?;
+    /// Assemble the index from the caller-constructed inverted index and the
+    /// finalization fields resolved from a [`SparseOpenPlan`]. Freshly (re)built
+    /// indexes are persisted (`persist`); loaded ones are not.
+    #[allow(clippy::too_many_arguments)]
+    pub fn finish(
+        config: SparseIndexConfig,
+        id_tracker: Arc<AtomicRefCell<IdTrackerEnum>>,
+        vector_storage: Arc<AtomicRefCell<VectorStorageEnum>>,
+        payload_index: Arc<AtomicRefCell<StructPayloadIndex>>,
+        path: &Path,
+        inverted_index: TInvertedIndex,
+        indices_tracker: IndicesTracker,
+        persist: bool,
+    ) -> OperationResult<Self> {
+        if persist {
+            config.save(&SparseIndexConfig::get_config_path(path))?;
+            inverted_index.save(path)?;
+            indices_tracker.save(path)?;
+            // Save the version last to mark a successful (re)build.
+            TInvertedIndex::Version::save(path)?;
+        }
 
-                config.save(&config_path)?;
-                inverted_index.save(path)?;
-                indices_tracker.save(path)?;
-
-                // Save the version as the last step to mark a successful rebuild.
-                TInvertedIndex::Version::save(path)?;
-
-                OperationResult::Ok((config, inverted_index, indices_tracker))
-            })?
-        };
-
-        let searches_telemetry = SparseSearchesTelemetry::new();
-        let path = path.to_path_buf();
-        let search_scratch_pool = SearchScratchPool::new();
         Ok(Self {
             config,
             id_tracker,
             vector_storage,
             payload_index,
-            path,
+            path: path.to_path_buf(),
             inverted_index,
-            searches_telemetry,
+            searches_telemetry: SparseSearchesTelemetry::new(),
             indices_tracker,
-            search_scratch_pool,
+            search_scratch_pool: SearchScratchPool::new(),
         })
     }
 
-    fn try_load(
-        path: &Path,
-    ) -> OperationResult<(SparseIndexConfig, TInvertedIndex, IndicesTracker)> {
-        let stored_version = TInvertedIndex::Version::load_universal(&MmapFs, path)?;
-
-        if stored_version != Some(TInvertedIndex::Version::current()) {
-            return Err(OperationError::service_error_light(format!(
-                "Index version mismatch, expected {}, found {}",
-                TInvertedIndex::Version::current(),
-                stored_version.map_or_else(|| "none".to_string(), |v| v.to_string()),
-            )));
-        }
-
-        let loaded_config = SparseIndexConfig::load(&SparseIndexConfig::get_config_path(path))?;
-        let inverted_index = TInvertedIndex::open(path)?;
-        let indices_tracker = IndicesTracker::open(path)?;
-        Ok((loaded_config, inverted_index, indices_tracker))
-    }
-
-    fn build_inverted_index(
+    /// Build the (type-agnostic) RAM inverted index and indices tracker from the
+    /// vector storage.
+    fn build_ram_index(
         id_tracker: &AtomicRefCell<IdTrackerEnum>,
         vector_storage: &AtomicRefCell<VectorStorageEnum>,
-        path: &Path,
         stopped: &AtomicBool,
         mut tick_progress: impl FnMut(),
-    ) -> OperationResult<(TInvertedIndex, IndicesTracker)> {
+    ) -> OperationResult<(InvertedIndexRam, IndicesTracker)> {
         let borrowed_vector_storage = vector_storage.borrow();
         let borrowed_id_tracker = id_tracker.borrow();
         let deleted_bitslice = borrowed_vector_storage.deleted_vector_bitslice();
@@ -211,10 +238,7 @@ impl<TInvertedIndex: InvertedIndex> SparseVectorIndex<TInvertedIndex> {
             }
             tick_progress();
         }
-        Ok((
-            TInvertedIndex::from_ram_index(Cow::Owned(ram_index_builder.build()), path)?,
-            indices_tracker,
-        ))
+        Ok((ram_index_builder.build(), indices_tracker))
     }
 
     pub fn inverted_index(&self) -> &TInvertedIndex {
