+++ lib/sparse/benches/search.rs
@@ -4,7 +4,7 @@ use std::sync::atomic::AtomicBool;
 
 use common::counter::hardware_counter::HardwareCounterCell;
 use common::types::PointOffsetType;
-use common::universal_io::MmapFile;
+use common::universal_io::{MmapFile, MmapFs};
 use criterion::measurement::Measurement;
 use criterion::{Criterion, criterion_group, criterion_main};
 use dataset::Dataset;
@@ -156,14 +156,15 @@ pub fn run_bench(
 
             run_bench2(
                 c.benchmark_group(format!("search/ram_{}/{name}", $name)),
-                &InvertedIndexCompressedImmutableRam::<$type>::open(&index_path).unwrap(),
+                &InvertedIndexCompressedImmutableRam::<$type>::open(&MmapFs, &index_path).unwrap(),
                 query_vectors,
                 &hottest_query_vectors,
             );
 
             run_bench2(
                 c.benchmark_group(format!("search/mmap_{}/{name}", $name)),
-                &InvertedIndexCompressedMmap::<$type, MmapFile>::open(&index_path).unwrap(),
+                &InvertedIndexCompressedMmap::<$type, MmapFile>::open(&MmapFs, &index_path)
+                    .unwrap(),
                 query_vectors,
                 &hottest_query_vectors,
             );
@@ -270,8 +271,12 @@ fn cached_compressed_index<W: Weight>(index: &InvertedIndexRam, name: &str) -> P
             fs::remove_dir_all(&tmp_path).unwrap();
         }
         fs::create_dir_all(&tmp_path).unwrap();
-        InvertedIndexCompressedMmap::<W, MmapFile>::from_ram_index(Cow::Borrowed(index), &tmp_path)
-            .unwrap();
+        InvertedIndexCompressedMmap::<W, MmapFile>::from_ram_index(
+            &MmapFs,
+            Cow::Borrowed(index),
+            &tmp_path,
+        )
+        .unwrap();
         fs::rename(tmp_path, &path).unwrap();
     }
     eprintln!("Using cache: {path:?}");
